feat: implement category-wise spending chart and synchronized analytics interactions
- Add a "Categories" view mode to `BudgetGraph` featuring a stacked bar chart (`MultiCategoryChart`) alongside the existing "Trend" cumulative view.
- Implement bidirectional synchronization between `BudgetGraph` and `CalendarHeatmap`, highlighting the chart's visible date range on the calendar.
- Update `CalendarHeatmap` interactions to focus the selected date on the detailed chart instead of opening a bottom sheet.
- Enhance `DetailedChart` with interactive tooltips that resolve category names and support external date selection.
- Add database migration (v15 to v16) to back-fill missing `categoryId` values by matching transaction comments against existing categories.
- Update `MinusCsvService` to automatically resolve or create categories during CSV imports based on transaction comments.
- Refactor `AnalyticsViewModel` to include hidden categories in the display state, ensuring historical transactions show correct labels even if the category was later deleted or hidden.
- Modularize `BudgetGraph` components into dedicated files for headers, navigation, and specific chart types.
- Add comprehensive screenshot and unit tests for category view modes, tooltips, and synchronized highlighting.
