+++ app/src/test/java/com/serranoie/app/minus/presentation/ui/screenshot/BudgetGraphScreenshotTest.kt
@@ -7,16 +7,19 @@ import androidx.compose.ui.unit.dp
 import app.cash.paparazzi.DeviceConfig
 import app.cash.paparazzi.Paparazzi
 import com.android.ide.common.rendering.api.SessionParams
+import com.serranoie.app.minus.domain.model.BudgetPeriod
+import com.serranoie.app.minus.domain.model.BudgetSettings
+import com.serranoie.app.minus.domain.model.Category
 import com.serranoie.app.minus.domain.model.Transaction
 import com.serranoie.app.minus.presentation.ui.analytics.AnalyticsState
 import com.serranoie.app.minus.presentation.ui.analytics.GraphGranularity
 import com.serranoie.app.minus.presentation.ui.theme.MinusTheme
 import com.serranoie.app.minus.presentation.ui.theme.component.budget.BudgetGraph
+import com.serranoie.app.minus.presentation.ui.theme.component.budget.BudgetGraphViewMode
 import org.junit.Rule
 import org.junit.Test
 import java.math.BigDecimal
 import java.time.LocalDate
-import java.time.LocalDateTime
 import java.time.ZoneId
 import java.util.Date
 import java.util.Locale
@@ -33,6 +36,14 @@ class BudgetGraphScreenshotTest {
         return Date.from(LocalDate.of(year, month, day).atStartOfDay(ZoneId.systemDefault()).toInstant())
     }
 
+    private fun sampleSettings(start: LocalDate, end: LocalDate) = BudgetSettings(
+        totalBudget = BigDecimal("1000"),
+        period = BudgetPeriod.MONTHLY,
+        startDate = start,
+        endDate = end,
+        currencyCode = "USD",
+    )
+
     private fun mockTransactions(start: LocalDate): List<Transaction> {
         return listOf(
             Transaction(amount = BigDecimal("150.00"), date = start.atTime(10, 0)),
@@ -43,9 +54,10 @@ class BudgetGraphScreenshotTest {
     }
 
     @Test
-    fun budgetGraph_daysView() {
+    fun budgetGraph_daysView_cumulative() {
         Locale.setDefault(Locale.US)
         val start = LocalDate.of(2026, 8, 1)
+        val end = LocalDate.of(2026, 8, 30)
         paparazzi.snapshot {
             MinusTheme {
                 Box(modifier = Modifier.padding(16.dp)) {
@@ -54,10 +66,109 @@ class BudgetGraphScreenshotTest {
                             spends = mockTransactions(start),
                             startPeriodDate = fixedDate(2026, 8, 1),
                             finishPeriodDate = fixedDate(2026, 8, 30),
+                            budgetSettingsForDisplay = sampleSettings(start, end),
                             graphGranularity = GraphGranularity.DAYS,
                             currencyCode = "USD"
                         ),
-                        onGranularityChanged = {}
+                        onGranularityChanged = {},
+                    )
+                }
+            }
+        }
+    }
+
+    @Test
+    fun budgetGraph_categoriesView() {
+        Locale.setDefault(Locale.US)
+        val start = LocalDate.of(2026, 8, 1)
+        val end = LocalDate.of(2026, 8, 7)
+        val categories = listOf(
+            Category(id = 1L, name = "Groceries"),
+            Category(id = 2L, name = "Transport"),
+        )
+        val spends = listOf(
+            Transaction(amount = BigDecimal("150.00"), date = start.atTime(10, 0), categoryId = 1L),
+            Transaction(amount = BigDecimal("85.50"), date = start.plusDays(2).atTime(10, 0), categoryId = 2L),
+            Transaction(amount = BigDecimal("60.00"), date = start.plusDays(2).atTime(18, 0), categoryId = 1L),
+            Transaction(amount = BigDecimal("120.00"), date = start.plusDays(5).atTime(10, 0), categoryId = null),
+        )
+        paparazzi.snapshot {
+            MinusTheme {
+                Box(modifier = Modifier.padding(16.dp)) {
+                    BudgetGraph(
+                        state = AnalyticsState(
+                            spends = spends,
+                            categories = categories,
+                            startPeriodDate = fixedDate(2026, 8, 1),
+                            finishPeriodDate = fixedDate(2026, 8, 7),
+                            budgetSettingsForDisplay = sampleSettings(start, end),
+                            graphGranularity = GraphGranularity.DAYS,
+                            currencyCode = "USD"
+                        ),
+                        onGranularityChanged = {},
+                        initialViewMode = BudgetGraphViewMode.CATEGORIES,
+                    )
+                }
+            }
+        }
+    }
+
+    @Test
+    fun budgetGraph_categoriesView_selectedDayHighlight() {
+        Locale.setDefault(Locale.US)
+        val start = LocalDate.of(2026, 8, 1)
+        val end = LocalDate.of(2026, 8, 7)
+        val categories = listOf(
+            Category(id = 1L, name = "Groceries"),
+            Category(id = 2L, name = "Transport"),
+        )
+        val spends = listOf(
+            Transaction(amount = BigDecimal("150.00"), date = start.atTime(10, 0), categoryId = 1L),
+            Transaction(amount = BigDecimal("85.50"), date = start.plusDays(2).atTime(10, 0), categoryId = 2L),
+            Transaction(amount = BigDecimal("60.00"), date = start.plusDays(2).atTime(18, 0), categoryId = 1L),
+            Transaction(amount = BigDecimal("120.00"), date = start.plusDays(5).atTime(10, 0), categoryId = null),
+        )
+        paparazzi.snapshot {
+            MinusTheme {
+                Box(modifier = Modifier.padding(16.dp)) {
+                    BudgetGraph(
+                        state = AnalyticsState(
+                            spends = spends,
+                            categories = categories,
+                            startPeriodDate = fixedDate(2026, 8, 1),
+                            finishPeriodDate = fixedDate(2026, 8, 7),
+                            budgetSettingsForDisplay = sampleSettings(start, end),
+                            graphGranularity = GraphGranularity.DAYS,
+                            currencyCode = "USD"
+                        ),
+                        onGranularityChanged = {},
+                        initialViewMode = BudgetGraphViewMode.CATEGORIES,
+                        selectedDate = start.plusDays(2),
+                    )
+                }
+            }
+        }
+    }
+
+    @Test
+    fun budgetGraph_categoriesView_empty() {
+        Locale.setDefault(Locale.US)
+        val start = LocalDate.of(2026, 8, 1)
+        val end = LocalDate.of(2026, 8, 7)
+        paparazzi.snapshot {
+            MinusTheme {
+                Box(modifier = Modifier.padding(16.dp)) {
+                    BudgetGraph(
+                        state = AnalyticsState(
+                            spends = emptyList(),
+                            startPeriodDate = fixedDate(2026, 8, 1),
+                            finishPeriodDate = fixedDate(2026, 8, 7),
+                            budgetSettingsForDisplay = sampleSettings(start, end),
+                            graphGranularity = GraphGranularity.DAYS,
+                            currencyCode = "USD"
+                        ),
+                        onGranularityChanged = {},
+                        initialViewMode = BudgetGraphViewMode.CATEGORIES,
                     )
                 }
             }
@@ -68,6 +179,7 @@ class BudgetGraphScreenshotTest {
     fun budgetGraph_totalView() {
         Locale.setDefault(Locale.US)
         val start = LocalDate.of(2026, 8, 1)
+        val end = LocalDate.of(2026, 8, 30)
         paparazzi.snapshot {
             MinusTheme {
                 Box(modifier = Modifier.padding(16.dp)) {
@@ -76,6 +188,7 @@ class BudgetGraphScreenshotTest {
                             spends = mockTransactions(start),
                             startPeriodDate = fixedDate(2026, 8, 1),
                             finishPeriodDate = fixedDate(2026, 8, 30),
+                            budgetSettingsForDisplay = sampleSettings(start, end),
                             graphGranularity = GraphGranularity.TOTAL,
                             currencyCode = "USD"
                         ),
