+++ app/src/test/java/com/serranoie/app/minus/presentation/ui/screenshot/DetailedChartScreenshotTest.kt
@@ -5,16 +5,19 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.unit.dp
 import app.cash.paparazzi.DeviceConfig
 import app.cash.paparazzi.Paparazzi
 import com.android.ide.common.rendering.api.SessionParams
+import com.serranoie.app.minus.domain.model.Category
 import com.serranoie.app.minus.domain.model.Transaction
 import com.serranoie.app.minus.presentation.ui.theme.MinusTheme
 import com.serranoie.app.minus.presentation.ui.theme.component.charts.DetailedChart
 import org.junit.Rule
 import org.junit.Test
 import java.math.BigDecimal
+import java.time.LocalDate
 import java.time.LocalDateTime
 import java.util.Locale
 
@@ -132,15 +135,74 @@ class DetailedChartScreenshotTest {
         }
     }
 
+    @Test
+    fun detailedChart_tooltip() {
+        Locale.setDefault(Locale.US)
+        val categories = listOf(
+            Category(id = 1L, name = "Groceries"),
+            Category(id = 2L, name = "Transport"),
+        )
+        paparazzi.snapshot {
+            MinusTheme {
+                Surface(modifier = Modifier.fillMaxWidth()) {
+                    DetailedChart(
+                        spends = listOf(
+                            sampleTransaction(id = 1L, amount = "100.00", daysAgo = 7, categoryId = 1L),
+                            sampleTransaction(id = 2L, amount = "60.00", daysAgo = 5, categoryId = 2L),
+                            sampleTransaction(id = 3L, amount = "180.00", daysAgo = 3, categoryId = 1L),
+                            sampleTransaction(id = 4L, amount = "75.00", daysAgo = 0, categoryId = null),
+                        ),
+                        categories = categories,
+                        onPointClick = {},
+                        forcedTouchPosition = Offset(280f, 150f),
+                        modifier = Modifier
+                            .fillMaxWidth()
+                            .height(250.dp),
+                    )
+                }
+            }
+        }
+    }
+
+    @Test
+    fun detailedChart_selectedDate() {
+        Locale.setDefault(Locale.US)
+        val categories = listOf(
+            Category(id = 1L, name = "Groceries"),
+            Category(id = 2L, name = "Transport"),
+        )
+        paparazzi.snapshot {
+            MinusTheme {
+                Surface(modifier = Modifier.fillMaxWidth()) {
+                    DetailedChart(
+                        spends = listOf(
+                            sampleTransaction(id = 1L, amount = "100.00", daysAgo = 7, categoryId = 1L),
+                            sampleTransaction(id = 2L, amount = "60.00", daysAgo = 5, categoryId = 2L),
+                            sampleTransaction(id = 3L, amount = "180.00", daysAgo = 3, categoryId = 1L),
+                            sampleTransaction(id = 4L, amount = "75.00", daysAgo = 0, categoryId = null),
+                        ),
+                        categories = categories,
+                        selectedDate = LocalDate.of(2026, 1, 15).minusDays(5),
+                        modifier = Modifier
+                            .fillMaxWidth()
+                            .height(250.dp),
+                    )
+                }
+            }
+        }
+    }
+
     private fun sampleTransaction(
         id: Long,
         amount: String,
         daysAgo: Long,
+        categoryId: Long? = null,
     ): Transaction = Transaction(
         id = id,
         amount = BigDecimal(amount),
         comment = "Sample",
         date = LocalDateTime.of(2026, 1, 15, 12, 0).minusDays(daysAgo),
         periodId = 7L,
+        categoryId = categoryId,
     )
 }
